working on cloud stuff

+++ scanbot_interface.py
@@ -541,10 +541,14 @@ class scanbot_interface(object):
         return args
     
     def uploadToCloud(self,filename):
+        print(filename)
+        if(not filename.endswith(".pkl")): return filename
         try:
+            print("Uploading to cloud")
             subprocess.run(["scp", filename, self.cloudPath])
             os.remove(filename)
         except Exception as e:
+            print("Failed")
             self.sendReply("Error uploading file to cloud with command\nscp " +
                            filename + " " + self.cloudPath + "\n\n" + str(e))
